JWT auth and bcrypt hashing addedd

+++ appCentralService/app.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const reload = require('reload');
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcryptjs');
 const app = express();
 const io = require('socket.io')();
 require('dotenv').config({path: `${__dirname}/.appcs.test.env`});
@@ -24,15 +26,84 @@ const server = app.listen(app.get('port'));
 console.log(`app listening on ${app.get('port')}`);
 io.attach(server);
 
-//TODO: register route. make a register route that: accepts
-//TODO: a person's username, and sends them back a response that authorizes
-//TODO: them to establish a socket connection. (with a JWT token)
-//TODO: cool shit man. this would be fun.
+//TODOs: General todos
+/*
 
-//TODO: make a login route that: accepts
-//TODO: a person's username, pw and sends them back a response that authorizes
-//TODO: them to establish a socket connection. (with a JWT token)
+1. Authify the important routes
 
+
+
+ */
+
+
+/*
+POST: /appcs/game/user/new.
+register a new user
+
+POST BODY:
+username
+password
+ */
+// there should be some client side validation on NON NULL name.
+app.post('/appcs/user/new',(req,res)=>{
+    db.getUser(req.body.username).then((user)=>{
+        if(!user) {
+            const salt = 10;
+            const userObj ={
+                username: req.body.username,
+             password: bcrypt.hashSync(req.body.password,salt)
+            };
+            db.registerUser(userObj).then(() => {
+                let token = jwt.sign({username:userObj.username,password:userObj.password},
+                    process.env.AUTH_TOKEN_SECRET,{
+                        expiresIn: 43200 // secs, so 12 hours.
+                    });
+                res.status(200).json({
+                    success:true,
+                    token: token
+                })
+            }).catch((e) => {
+                res.json({
+                    success: false
+                });
+            })
+        } else
+            // a user is already defined with that name
+            res.json({
+                success:false,
+                error: 'User already exists.'
+            });
+    })
+});
+
+/*
+
+POST /appcs/user/auth
+// Authenticates user (login)
+POST BODY:
+// username
+// password
+
+ */
+app.post('/appcs/user/auth',(req,res)=>{
+    db.getUser(req.body.user).then((user)=>{
+        const passwordValid = bcrypt.compareSync(req.body.password,user.password);
+        if(passwordValid){
+            let token = jwt.sign({username:user.username,password:user.password},
+                process.env.AUTH_TOKEN_SECRET,
+                {expiresIn: 43200});
+            res.status(200).send({
+                success:true,
+                token: token
+            });
+        }else{
+            res.status(401).send({
+                success:false,
+                token: null
+            })
+        }
+    })
+})
 
 
 
@@ -67,29 +138,37 @@ AppCS Route.
 
 POST body:
 
-    game: {
-        // game object.
-    }
+    game: game object
+    token
 
 TESTED. OK.
 
  */
 app.post('/appcs/game/create', (req, res) => {
     // the creator information is here already. (inside req.body.game object).
-    db.createGame(JSON.parse(req.body.game)).then((newgame) => {
-        // link used to go to the lobby page.
-        io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_CREATED, {
-            game: newgame
+    jwt.verify(req.body.token,process.env.AUTH_TOKEN_SECRET,(err,decoded)=>{
+        if(err)
+            res.json({
+                success:false,
+                error: 'NOT AUTHENTICATED.'
+            });
+        let game = req.body.game;
+        game.creator = decoded.username;
+        db.createGame(JSON.parse(game)).then((newgame) => {
+            // link used to go to the lobby page.
+            io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_CREATED, {
+                game: newgame
+            });
+            res.json({
+                success: true,
+                game: newgame
+            })
+        }).catch((e) => {
+            res.json({
+                success: false,
+            })
         });
-        res.json({
-            success: true,
-            game: newgame
-        })
-    }).catch((e) => {
-        res.json({
-            success: false,
-        })
-    });
+    })
 });
 
 
@@ -101,30 +180,37 @@ DELETE /appcs/game/delete/:gameId: Delete game
 POST body:
 
     socketid: "awdijaos2123eda",
+    token
 
 TESTED. OK.
  */
 app.delete('/appcs/game/delete/:gameid', (req, res) => {
-    db.getGame(req.params.gameid).then((game) => {
-        db.getUserSecrets(game.creator).then((creator)=>{
-            if (creator.socketid === req.body.socketid) {
-                db.deleteGame(req.params.gameid).then(() => {
-                    io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_DELETED, {
-                        gameid: req.params.gameid
+    jwt.verify(req.body.token,process.env.AUTH_TOKEN_SECRET,(err,decoded)=> {
+        if(err)
+            res.json({success:false, error:'unauthorized.'});
+
+        db.getGame(req.params.gameid).then((game) => {
+            db.getUserSecrets(game.creator).then((creator) => {
+                if (creator.socketid === req.body.socketid &&
+                creator.username === decoded.username) {
+                    db.deleteGame(req.params.gameid).then(() => {
+                        io.of(MAIN_NAMESPACE).emit(EVENTS.GAME_DELETED, {
+                            gameid: req.params.gameid
+                        });
+                        io.of(MAIN_NAMESPACE)
+                            .to(req.params.gameId)
+                            .emit(EVENTS.LOBBY.LOBBY_GAME_DELETED);
+                        res.json({
+                            success: true,
+                        })
+                    }).catch((e) => {
+                        res.json({
+                            success: false,
+                        })
                     });
-                    io.of(MAIN_NAMESPACE)
-                        .to(req.params.gameId)
-                        .emit(EVENTS.LOBBY.LOBBY_GAME_DELETED);
-                    res.json({
-                        success: true,
-                    })
-                }).catch((e) => {
-                    res.json({
-                        success: false,
-                    })
-                });
-            }
-        })
+                }
+            })
+        });
     });
 });
 
@@ -135,39 +221,62 @@ AppCS Route.
 POST /appcs/game/start/:gameId: Start game
 NOTE: we communicate with our GMS here.
 POST body:
-{
-    gameId:"awdnsoawd", // game id of one to be deleted.
-    socketId:"ajwodmala" // socket id of requester
-}
+
+    gameid:"awdnsoawd", // game id of one to be deleted.
+    socketid:"ajwodmala" // socket id of requester
+    token
+
  */
 
 app.post('/appcs/game/start/:gameid', (req, res) => {
-    db.getGame(req.params.gameid).then((game) => {
-        db.getUserSecrets(game.creator).then((creator) => {
-            if (creator.socketid === req.body.socketid) {
-                io.of(MAIN_NAMESPACE)
-                    .to(req.params.gameid)
-                    .emit(EVENTS.LOBBY.GAME_START);
-                //TODO
-                //TODO: communicate with GMS.
-                res.json({
-                    success: true
-                });
-            }
-            else
-                res.json({
-                    success: false
-                });
+    jwt.verify(req.body.token, process.env.AUTH_TOKEN_SECRET,(err,decoded)=>{
+        if(err)
+            res.json({
+                success:false,
+                error: 'Unauthorized.'
+            });
+        db.getGame(req.params.gameid).then((game) => {
+            db.getUserSecrets(game.creator).then((creator) => {
+                if (creator.socketid === req.body.socketid) {
+                    io.of(MAIN_NAMESPACE)
+                        .to(req.params.gameid)
+                        .emit(EVENTS.LOBBY.GAME_START);
+                    //TODO
+                    //TODO: communicate with GMS.
+                    res.json({
+                        success: true
+                    });
+                }
+                else
+                    res.json({
+                        success: false
+                    });
+            });
         });
-    });
+    })
 });
+// WS routes: authenticated
+// we use our middleware to deal with JWT auth
+io.use(function(socket,next){
+    if(socket.handshake.token){
+        jwt.verify(socket.handshake.token, process.env.AUTH_TOKEN_SECRET,(err,decoded)=>{
+            if(err)
+                return next(new Error('WS Auth Error'));
 
-io.of(MAIN_NAMESPACE).on('connect', (socket) => {
+            socket.username = decoded.username;
+            next();
+        })
+    }else{
+        next(new Error('WS Authentication Error'));
+    }
+}).of(MAIN_NAMESPACE).on('connect', (socket) => {
     if (!socket.sentMydata) {
-        socket.emit(EVENTS.CONN_DETAILS, {socketId: socket.id});
+        //NOTE: I don't know if this will work or not.
+        const asyn = async ()=>await db.loginUserSocketId(socket.username, socket.id);
+        asyn();
         socket.sentMydata = true;
     }
-
+    /*
     /*
 
     AppCS Route.
@@ -177,7 +286,7 @@ io.of(MAIN_NAMESPACE).on('connect', (socket) => {
     socket.on(EVENTS.LOBBY.CLIENT_ATTEMPT_JOIN, (clientUserObj, gameId) => {
         let roomName = gameId;
         // Could be a hash.
-        db.joinGame(clientUserObj.username, gameId).then(() => {
+        db.joinGame(clientUserObj.username, roomName).then(() => {
             socket.join(roomName);
             clientUserObj.socketid= null;
             io.of(MAIN_NAMESPACE)
